Rename files to use UpperCamelCaseCluster/Logic for user-label-server
