+++ src/app/clusters/user-label-server/BUILD.gn
@@ -16,8 +16,8 @@ import("//build_overrides/chip.gni")
 
 source_set("user-label-server") {
   sources = [
-    "user-label-cluster.cpp",
-    "user-label-cluster.h",
+    "UserLabelCluster.cpp",
+    "UserLabelCluster.h",
   ]
 
   public_deps = [
